исправляет мелкие недочеты и переносит константы в отдельный файл
