+++ js/api.js
@@ -1,19 +1,4 @@
-const BASE_URL = 'https://32.javascript.htmlacademy.pro/kekstagram';
-
-const Route = {
-  GET_DATA: '/data',
-  SEND_DATA: '/',
-};
-
-const Method = {
-  GET: 'GET',
-  POST: 'POST',
-};
-
-const ErrorText = {
-  [Method.GET]: 'Не удалось загрузить данные. Попробуйте обновить страницу',
-  [Method.POST]: 'Не удалось отправить форму. Попробуйте ещё раз',
-};
+import { BASE_URL, Route, Method, ErrorText } from './const.js';
 
 const load = async (route, method = Method.GET, body = null) => {
   const response = await fetch(`${BASE_URL}${route}`, {method, body});
